serializeファイルの存在を確認して振る舞うように修正

+++ src/test/java/jp/kusumotolab/kgenprog/project/test/TestProcessBuilderTest.java
@@ -17,7 +17,9 @@ public class TestProcessBuilderTest {
 
 	@Before
 	public void before() throws IOException {
-		Files.delete(TestResults.getSerFilePath());
+		if (Files.exists(TestResults.getSerFilePath())) {
+			Files.delete(TestResults.getSerFilePath());
+		}
 	}
 
 	@Test
